move collectFields out of generated code
