+++ neelance/errors/errors.go
@@ -2,8 +2,6 @@ package errors
 
 import (
 	"fmt"
-
-	"github.com/vektah/gqlgen/graphql"
 )
 
 type QueryError struct {
@@ -52,35 +50,4 @@ func (c *Builder) Errorf(format string, args ...interface{}) {
 
 func (c *Builder) Error(err error) {
 	c.Errors = append(c.Errors, Errorf("%s", err.Error()))
-}
-
-func ErrorWriter(errs []*QueryError) graphql.Marshaler {
-	res := graphql.Array{}
-
-	for _, err := range errs {
-		if err == nil {
-			res = append(res, graphql.Null)
-			continue
-		}
-
-		errObj := &graphql.OrderedMap{}
-
-		errObj.Add("message", graphql.String(err.Message))
-
-		if len(err.Locations) > 0 {
-			locations := graphql.Array{}
-			for _, location := range err.Locations {
-				locationObj := &graphql.OrderedMap{}
-				locationObj.Add("line", graphql.Int(location.Line))
-				locationObj.Add("column", graphql.Int(location.Column))
-
-				locations = append(locations, locationObj)
-			}
-
-			errObj.Add("locations", locations)
-		}
-		res = append(res, errObj)
-	}
-
-	return res
 }
